Add `style` option
This adds a `style` option with the possible values: `"best fit"` (default), and `"numeric"`.
When the `{style: "numeric"}` is set, the formatted output will _always_ contain a number. Whereas, when the default `{style: "best fit"}` is used, relative units (e.g., "yesterday") might be chosen as the best output for the given locale and relative date.

Tests have also been added for both `options`.